Refactored into delegates in LoadingViewController, calling Bart API
